Add Render that doesn't return error; deprecate Must(RenderNodes()).
This is a somewhat experimental change and may be reverted if it
doesn't work well in practice.

The motivation is that in most (all) uses of htmlg, the callers end up
doing panic(err) or using htmlg.Must (which does the same). This is
because it's never expected to fail on the inputs it's provided, so
there's no point in another type of error handling. html.Render may
return errors when the supplied io.Writer fails, but that shouldn't
happen with a bytes.Buffer unless we're out of memory, where panic is
the most appropriate response.

Therefore, make the API nicer by removing error return type.