Add ComponentContext and RenderComponentsContext.
This is very experimental, as I'm not sure if it's a good idea to be
doing things that require a context inside Render. Having a context
implies doing something potentially slow, that potentially may return
non-nil error. How to handle that in Render which doesn't return error?
In theory, the component can render itself differently to display an
error, but is that a good idea overall?

Maybe such things should be done by a higher level component, like a
store, etc.

We will try it and find out. Clearly, these are experimental and may
change or go away.