Deprecate ComponentContext, RenderComponentsContext.
They were an experiment, and it turned out not to be a good idea. It
makes more sense and scales better to perform service requests in
advance (potentially in parallel) and pass static values to components
to render.

Sometimes a part or superset of required information is already
available, so it's better to provide it directly to the component
to use, rather than have it make the same (or similar) service request
again, or need to write a fake service implementation that instantly
returns available data.

It also makes components more unpredictable in terms of performance.

Updates 4140b7e77c3d62312d57f95600cca22344258a54.