add Nodes type which implements Component interface from []*html.Node
This is handy when one wants to provide a list of HTML node elements
into something that requires an html.Component. For example:

	err := htmlg.RenderComponents(w, component.Flash{
		Content: htmlg.Nodes{
			htmlg.Text("Showing partial commit. Click "),
			htmlg.A("here", "/foo/bar"),
			htmlg.Text(" for full commit."),
		},
	})
	if err != nil {
		// handle error
	}

This implementation of the Component interface is unusual in that
the Render method always returns the same references to existing
nodes, and as a result, it is unsuitable to be rendered and attached
to other HTML nodes more than once. It is suitable to rendered
directly into HTML multiple times, or attached to an existing node once.

Add it for now in order to gain experience using it and iterate
towards better ideas, but document it as experimental and able
to change or get removed.